+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: (EVM) Added exponential backoff function to etherscan/evmscan API call retries.
+
 ## 4.51.0 (2025-06-24)
 
 - changed: Osmosis archive node urls